Renamed FeatureGate as RuntimeOptions
