+++ vnext/Desktop/RuntimeOptions.cpp
@@ -1,4 +1,4 @@
-#include <FeatureGate.h>
+#include <RuntimeOptions.h>
 
 #include <mutex>
 #include <unordered_map>
