+++ vnext/Shared/RuntimeOptions.h
